adding routers for api, articles, comments & topics. this also includes a refactor to move the article/:article_id/comments endpoins into the articles model & controller to allow the usage of routing
