+++ app/routes/topics-router.js
@@ -1,0 +1,6 @@
+const topicsRouter = require('express').Router();
+const { getTopics } = require('../../controllers/topics.controllers');
+
+topicsRouter.get('/', getTopics);
+
+module.exports = topicsRouter;
