+++ controllers/articles.contoller.js
@@ -2,6 +2,8 @@ const {
   fetchArticleById,
   fetchArticles,
   updateArticleVotes,
+  fetchCommentsByArticleId,
+  createNewComment,
 } = require('../models/articles.model');
 
 exports.getArticles = async (request, response, next) => {
@@ -35,6 +37,29 @@ exports.patchArticle = async (request, response, next) => {
       addOrRemoveVotes
     );
     response.status(200).send({ article: updatedArticle });
+  } catch (error) {
+    next(error);
+  }
+};
+
+exports.getCommentsByArticleId = async (request, response, next) => {
+  try {
+    const { article_id } = request.params;
+    const comments = await fetchCommentsByArticleId(article_id);
+    response.status(200).send({ comments });
+  } catch (error) {
+    next(error);
+  }
+};
+
+exports.postCommentbyArticleId = async (request, response, next) => {
+  try {
+    const {
+      params: { article_id },
+      body,
+    } = request;
+    const newComment = await createNewComment(article_id, body);
+    response.status(201).send({ comment: newComment });
   } catch (error) {
     next(error);
   }
