+++ package.json
@@ -9,7 +9,7 @@
     "test": "jest --runInBand",
     "prepare": "husky install",
     "test-seed": "jest seed.test.js",
-    "start": "node ./app/listen.js",
+    "start": "nodemon ./app/listen.js",
     "seed-prod": "NODE_ENV=production npm run seed-dev",
     "start-prod": "NODE_ENV=production npm run start"
   },
